+++ rules/rule_algo.py
@@ -44,37 +44,37 @@ def propertyDamage(field, housesDamaged, hotelDamaged):
     decision = random.randint(0, 1)
     
     if housesDamaged > 0:
-        if hotelDamaged:
+        if hotelDamaged and field.hotelAvailable:
             r_ui.latest_event = [f"{str(housesDamaged)} house(s) & the hotel in {field.name} damaged."]
             
         else:
             r_ui.latest_event = [f"{str(housesDamaged)} house(s) in {field.name} damaged."]
             
-    elif hotelDamaged:
+    elif hotelDamaged and field.hotelAvailable:
         r_ui.latest_event = [f"Hotel in {field.name} damaged."]
     
     if decision == 0:
         if field.houseCount > 0:
             field.houseCount -= housesDamaged
         
-        field.price /= housesDamaged
-        field.rent /= housesDamaged
+            field.price /= housesDamaged
+            field.rent /= housesDamaged
         
-        if hotelDamaged:
+        if hotelDamaged and field.hotelAvailable:
+            field.hotelAvailable = False
             field.price -= 40
             field.rent -= 4
             
-        r_ui.latest_event.append("Price 40 less. Rent 4 less.")
+        r_ui.latest_event.append("Price and rent decrease.")
     
     elif decision == 1:
         if field.owner != "Bank":
-            field.owner.balance -= 25 * housesDamaged # Can be changed
+            field.owner.balance -= house_price * housesDamaged # Can be changed
             
-            if hotelDamaged:
-                field.owner.balance -= 50
+            if hotelDamaged and field.hotelAvailable:
+                field.owner.balance -= hotel_price
                 
-        r_ui.latest_event.append("Repairs: 25 per house & 50 for")
-        r_ui.latest_event.append("hotel")
+        r_ui.latest_event.append("Repairs to be done")
     
 def shopOpens(field):
     field.price *= 1.5
@@ -205,7 +205,10 @@ def eventSelector(screen, jail, players, dices, jail_fid):
             if field.type == "street":
                 st_container.append(field)
         
-        propertyDamage(st_container[random.randint(0, len(st_container) - 1)], random.randint(1, 4), random.choice([True, False]))
+        street_chosen = st_container[random.randint(0, len(st_container) - 1)]
+        
+        if street_chosen.houseCount > 0 or street_chosen.hotelAvailable:
+            propertyDamage(street_chosen, random.randint(1, street_chosen.houseCount), random.choice([True, False]))
         
     elif eventSel == 4:
         st_container = []