+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This build is not intended for submission, but rather for the developers to keep
 - There is an effort in fixing "drunk rent crashes".
 - The game would now check for *some* winning conditions.
 - The winning conditions are random.
+- The property damage event has been fixed.
     
 ## Known issues